Move KnownFitting enum into fittings.known_fitting.
Keep the closed catalog vocabulary out of models.py so guide key churn stays localized.
